+++ user_guide_src/source/changelog.rst
@@ -160,6 +160,7 @@ Release Date: Not Released
    -  Added support for SQLite3 database driver.
    -  Added Interbase/Firebird database support via the *ibase* driver.
    -  Added ODBC support for ``create_database()``, ``drop_database()`` and ``drop_table()`` in :doc:`Database Forge <database/forge>`.
+   -  Added **save_queries** config setting in *config/database.php*. Default is ``TRUE``.
 
    -  :doc:`Query Builder <database/query_builder>` changes include: